Work on conversation manager and changes to database adapters

+++ src/com/tectria/imrek/util/IMrekChannelDbAdapter.java
@@ -38,12 +38,28 @@ public class IMrekChannelDbAdapter {
 	}
 	
 	public boolean removeChannel(String channel_name) {
-		// TODO: cascade delete to any messages stored for this channel
+		IMrekMessageDbAdapter messageAdapter = new IMrekMessageDbAdapter(this.context);
+		if (!messageAdapter.clearChannel(getChannelId(channel_name))) {
+			// return false if we can't clear the messages for the channel
+			// this likely means that the channel doesn't actually exist and something
+			// is messed up.
+			return false;
+		}
 		return database.delete(DATABASE_TABLE, "channel_name = ?", new String[]{channel_name}) > 0;
 	}
 	
 	public Cursor getChannels() {
 		return database.query(DATABASE_TABLE, null, null, null, null, null, null);
+	}
+	
+	public long getChannelId(String name) {
+		Cursor c = database.query(DATABASE_TABLE, new String[]{"_id"}, "channel_name = ? ", new String[]{name}, null, null, null);
+		if (!c.moveToFirst()) {
+			// if there's no results or an error, return -1
+			return -1;
+		}
+		
+		return c.getLong(0);
 	}
 	
 }
